+++ packages/imperative/src/config/cmd/secure/secure.handler.ts
@@ -92,22 +92,22 @@ export default class SecureHandler implements ICommandHandler {
         const authHandlerClass = ConfigAutoStore.findAuthHandlerForProfile(profilePath, this.params.arguments);
 
         if (authHandlerClass != null) {
-            const [promptParams, loginHandler] = authHandlerClass.getPromptParams();
-
-            if (promptParams.serviceDescription != null) {
-                this.params.response.console.log(`Logging in to ${promptParams.serviceDescription}`);
+            const api = authHandlerClass.getAuthHandlerApi();
+            if (api.promptParams.serviceDescription != null) {
+                this.params.response.console.log(`Logging in to ${api.promptParams.serviceDescription}`);
             }
 
             const profile = config.api.profiles.get(profilePath.replace(/profiles\./g, ""), false);
-            const sessCfg: ISession = await ConnectionPropsForSessCfg.addPropsOrPrompt({}, profile as ICommandArguments,
-                { parms: this.params, doPrompting: true, requestToken: true, ...promptParams });
+            const sessCfg: ISession = api.createSessCfg(profile as ICommandArguments);
+            const sessCfgWithCreds = await ConnectionPropsForSessCfg.addPropsOrPrompt(sessCfg, profile as ICommandArguments,
+                { parms: this.params, doPrompting: true, requestToken: true, ...api.promptParams });
             Logger.getAppLogger().info(`Fetching ${profile.tokenType} for ${propPath}`);
 
             try {
-                return await loginHandler(new Session(sessCfg));
+                return await api.sessionLogin(new Session(sessCfgWithCreds));
             } catch (error) {
                 throw new ImperativeError({
-                    msg: `Failed to fetch ${profile.tokenType} for ${propPath}`,
+                    msg: `Failed to fetch ${profile.tokenType} for ${propPath}: ${error.message}`,
                     causeErrors: error
                 });
             }